backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
scikit-learn>=1.4.0 